safari-600.7-branch/LayoutTests/webaudio/delaynode-maxdelaylimit-expected.txt

Tests attribute and maximum allowed delay of DelayNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.createDelay(180) threw exception Error: NotSupportedError: DOM Exception 9.
PASS context.createDelay(0) threw exception Error: NotSupportedError: DOM Exception 9.
PASS context.createDelay(-1) threw exception Error: NotSupportedError: DOM Exception 9.
PASS delay.delayTime.value === 0.5 is true
PASS Test signal was correctly delayed by 0.5 sec.
PASS successfullyParsed is true

TEST COMPLETE
